Add files via upload
The word to number contains two classes. One for conversion of words to number and the other for conversion of number to words. both classes have the capacity of conversion of both whole numbers and decimal numbers and any combination of the two in both word to number and number to word conversions.
The code is organized to ensure optimized memory and fast runtime of the code.